fixes from namespace move
